LayoutTests/imported/w3c/web-platform-tests/css/css-animations/animation-base-response-004-expected.txt


FAIL Base is responsive to font-affecting appearing via setKeyframes assert_equals: expected "15px" but got "10px"
